separate implementations into simple vs. focus
