+++ benches/four_banners.rs
@@ -1,4 +1,5 @@
-use aksim::{calculate, Banner, Float};
+use aksim::simple::{calculate as calculate_simple, Banner};
+use aksim::Float;
 use divan::{bench, black_box, AllocProfiler};
 
 #[global_allocator]
@@ -15,31 +16,27 @@ fn target_1() -> Float {
             target: 1,
             subrate: 0.35,
             bonus_pulls: 24,
-            has_focus: false,
         },
         Banner {
             target: 1,
             subrate: 0.5,
             bonus_pulls: 0,
-            has_focus: false,
         },
         Banner {
             target: 1,
             subrate: 0.2,
             bonus_pulls: 48,
-            has_focus: false,
         },
         Banner {
             target: 1,
             subrate: 0.4,
             bonus_pulls: 0,
-            has_focus: false,
         },
     ];
 
     let pulls = 400;
 
-    calculate(black_box(&banners), black_box(pulls))
+    calculate_simple(black_box(&banners), black_box(pulls))
 }
 
 #[bench]
@@ -49,29 +46,25 @@ fn target_10() -> Float {
             target: 10,
             subrate: 0.35,
             bonus_pulls: 240,
-            has_focus: false,
         },
         Banner {
             target: 10,
             subrate: 0.5,
             bonus_pulls: 0,
-            has_focus: false,
         },
         Banner {
             target: 10,
             subrate: 0.2,
             bonus_pulls: 480,
-            has_focus: false,
         },
         Banner {
             target: 10,
             subrate: 0.4,
             bonus_pulls: 0,
-            has_focus: false,
         },
     ];
 
     let pulls = 4000;
 
-    calculate(black_box(&banners), black_box(pulls))
+    calculate_simple(black_box(&banners), black_box(pulls))
 }
